+++ src/lib/services/verification/evm/index.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { CELATONE_QUERY_KEYS } from "lib/app-provider/env";
-import { getEvmVerifyConfig, getEvmVerifyInfo, submitEvmVerify } from "./api";
-import type { HexAddr20, Option } from "lib/types";
+import { getEvmVerifyConfig, getEvmVerifyInfos, submitEvmVerify } from "./api";
+import type { HexAddr20 } from "lib/types";
 import { useCurrentChain } from "lib/app-provider";
 import { isHex20Bytes } from "lib/utils/validate";
 
@@ -15,17 +15,22 @@ export const useEvmVerifyConfig = () =>
     retryOnMount: false,
   });
 
-export const useEvmVerifyInfo = (contractAddress: Option<HexAddr20>) => {
+export const useEvmVerifyInfos = (contractAddresses: HexAddr20[]) => {
   const { chainId } = useCurrentChain();
 
   return useQuery({
-    queryKey: [CELATONE_QUERY_KEYS.EVM_VERIFY_INFO, chainId, contractAddress],
+    queryKey: [
+      CELATONE_QUERY_KEYS.EVM_VERIFY_INFOS,
+      chainId,
+      contractAddresses,
+    ],
     queryFn: () => {
-      if (!contractAddress)
-        throw new Error("contractAddress is undefined (useEvmVerifyInfo)");
-      return getEvmVerifyInfo(chainId, contractAddress);
+      if (contractAddresses.length === 0)
+        throw new Error("contractAddresses is empty (useEvmVerifyInfo)");
+      return getEvmVerifyInfos(chainId, contractAddresses);
     },
-    enabled: !!contractAddress && isHex20Bytes(contractAddress),
+    enabled:
+      contractAddresses.length > 0 && contractAddresses.every(isHex20Bytes),
     refetchOnWindowFocus: false,
     retry: 1,
     staleTime: Infinity,