+++ CHANGELOG.md
@@ -83,6 +83,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1248](https://github.com/alleslabs/celatone-frontend/pull/1248) Support querying multiple evm verification infos
 - [#1246](https://github.com/alleslabs/celatone-frontend/pull/1246) Handle undefined constructor args
 - [#1242](https://github.com/alleslabs/celatone-frontend/pull/1242) Improve EVM proxy contract details and checksum address
 - [#1236](https://github.com/alleslabs/celatone-frontend/pull/1236) Remove unnecessary configs for users in custom network configs